feat: Flower 선물 Get 기능 구현 및 코드 리팩토링
1. Stack내부 구문이 길어진 경우 private func으로 따로 관리합니다.
2. FlowerList에 있던 일부 struct가 Model 폴더로 이전되었습니다.
3. FlowerView 파일 이름이 FlowerListView로 변경되었습니다.
4. FlowerList Get 기능이 구현되었습니다.
5. 개행 및 사소한 부분에서의 코드가 수정되었습니다.
