+++ liveOnReboot/liveOn/Network/Gift/Flower/DTO/FlowerDTO.swift
@@ -4,17 +4,14 @@
 //
 //  Created by Keum MinSeok on 2022/07/09.
 //
-
 import SwiftUI
 
-struct FlowerDTO: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
+struct FlowerGetResponse: Codable {
+    let giftFlowerId: Int
+    let giftFlowerImage: String
 }
 
-struct FlowerDTO_Previews: PreviewProvider {
-    static var previews: some View {
-        FlowerDTO()
-    }
+struct FlowerPostRequest: Codable {
+    let flowerId: String
+    let message: String
 }
