+++ liveOnReboot/liveOn/Network/Gift/Flower/EndPoint/FlowerEndPoint.swift
@@ -6,15 +6,53 @@
 //
 
 import SwiftUI
+import Moya
 
-struct FlowerEndPoint: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+extension FlowerServerCommunication: TargetType, AccessTokenAuthorizable {
+    public var baseURL: URL {
+        return URL(string: GeneralAPI.baseURL)!
     }
-}
 
-struct FlowerEndPoint_Previews: PreviewProvider {
-    static var previews: some View {
-        FlowerEndPoint()
+    var path: String {
+        switch self {
+            case .getFlower:
+                return "/api/v1/gifts/flowers"
+            case .postFlower:
+                return "/api/v1/gifts/flowers"
+        }
+    }
+
+    var method: Moya.Method {
+        switch self {
+            case .getFlower:
+                return .get
+            case .postFlower:
+                return .post
+        }
+    }
+
+    var task: Task {
+        switch self {
+            case .getFlower:
+                return .requestPlain
+            case .postFlower(let content):
+                return .requestJSONEncodable(content)
+        }
+    }
+
+    var authorizationType: AuthorizationType? {
+        switch self {
+
+        default:
+            return .bearer
+        }
+    }
+
+    var headers: [String : String]? {
+        switch self {
+            default:
+                return ["Content-Type": "application/json",
+                        "Authorization": "Bearer " + GeneralAPI.token]
+        }
     }
 }
