+++ liveOnReboot/liveOn/Network/Gift/Flower/Service/FlowerService.swift
@@ -4,17 +4,12 @@
 //
 //  Created by Keum MinSeok on 2022/07/09.
 //
-
 import SwiftUI
+import Moya
 
-struct FlowerService: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
+let flowerMoyaService = MoyaProvider<FlowerServerCommunication>(plugins: [NetworkLoggerPlugin()])
 
-struct FlowerService_Previews: PreviewProvider {
-    static var previews: some View {
-        FlowerService()
-    }
+enum FlowerServerCommunication {
+    case getFlower
+    case postFlower(content: FlowerPostRequest)
 }
